allow for unmatched node files
Prior to this change an error was reported if a node could not be
matched to a node file. This causes problems when a node file is
removed in a commit and heckler tries to find the owner for a node.

Instead allow for no matches so at least an admin could approve the
issue.