parentEvalErrors: Only note parent eval errors in noop issue
Prior to this commit the noop issue contained both a note stating that a
parentEvalError occurred as well as the full noop output for the child.
The full noop output is confusing to the user. Instead not that the
parent erroed, but do not show the full noop output of the child.

This commit does not add any notifications for these errors.