don't load invalid noops
Consider the following order of commits and applies

  0. Initial apply
  1. Edit file add 'A'
  2. Add comment A, no change (apply, after 3's noop is created)
  3. Edit file add 'B' - diff against apply (0)
    +A
    +B
  4. Add comment B, no change - (diff against (2), after (2) is applied)
    +B

If we subtract the noop from 3 from 4 the diffs will not match because
the noops were made against different applied revisions. To avoid this
problem consider any serialized noop invalid if its last apply does not
match the host's current last apply.