groupReportNodeSet: fix merge noops
Prior to this commit we subtracted parents from a merge commit even
if one of those parents was not in the lineage of all nodes. This
is incorrect as noops with are not in the lineage of all nodes are
not reliable. So ensure the parents are in all nodes lineage. The
by product of this change is that merge commits will now correctly
show a noop introduced by a branch.