heckler report: mark box as dirty on failed apply
Prior to this commit we did not create a last apply report when the
apply failed, however this meant partially applied resources from a
failed apply were not accounted for in the last apply status. Instead if
an apply fails mark the box as dirty, but still create a last apply
report.

This does overload the "dirty" suffix, but that might be ok, time will
tell.