reopen closed unapproved issues, notify when "commit authors" == 0
Prior to this commit unapproved issues that were closed, were not
reopened. This made it difficult to ascertain which issues still
needed approval. Also, when the commit authors' GitHub logins could
not be determined no indication was provided on the issue.

After this commit unapproved issues are reopened if they were closed.
Also the approval comment notes the need for admin approval when the
commit authors could not be determined.