noop: serialize the noop after creating the GitHub issue
Prior to this commit we created issues after serializing the grouped
report. However, this relied on checking if an existing issue existed.
Unfortunately, GitHub sometimes returns false negatives, which results
in duplicate issues. Instead create the issue prior to serialization and
if creating an issue fails, discard the grouped report.

We are now relying on the existence of a serialized grouped report to
indicate that an issue has been created.

The next step would be to used the serialized GitHub issue ID to
lookup issues.