Fix the make-repo script
It had a bunch of broken things in it, including git command flags that
weren't recognized, bad logic that made it try to copy files to places
that didn't exist yet or refer to branches that hadn't been defined, as
well as the general bad approach of init-ing a new repo within the
heckler repo. Instead, this does everything in the temp directory, and
won't accidentally mess your heckler clone up! The downside is that you
have to re-clone your test repo in a place you prefer to work from if
you want to iterate on it later, but that's small potatoes compared to
force-pushing literal broken changes onto your fork of heckler 😅

Co-authored-by: czavieh <czavieh@users.noreply.github.com>