Update integration test setup and instructions
This adds more details to the steps needed to even run the tests, plus
it makes them actually work. I decided that rather than having people
copy the sample configs I would make it so that they point to them with
the new -configfile argument, but I may change my mind about it later.
I'd like to get to the point where the integration tests have their own
config for the test app that we use to template the hecklerd and rizzod
configs, rather than making people copy and/or edit them.

Co-authored-by: Andrew Garrett <angarrett@paypal.com>