Define integration tests' success criteria somewhat concretely
We also decided to make the commands we ask the user to run a bit
simpler, by offloading the actually running of the daemons to systemd.
Since, for some reason, the foreground processes are not just the
daemons. I could not figure out a way to use build args to make two
images that run one or the other service, so unfortunately, I had to
switch the CMD to an ENTRYPOINT and make the service name an argument to
it. Really not a fan, but, whatever. At least now we are satisfied that
our changes thus far have not broken the app.

Co-authored-by: czavieh <czavieh@users.noreply.github.com>