Automate the integration tests partway
Setting up the app, the test repo, the conf file, and actually compiling
the binaries are still in their own steps, and we don't have very robust
things going on in the integration tests (or even a way to make it exit
non-0 if they fail) but this still eliminates a fair chunk of tedium.
More improvements to come.