Update go, libressl, musl, debian, and go mod versions, and make integration tests work (#18)
* Update go, libressl, musl, debian, and go mod versions

Had to fix the tests too. They wouldn't work when I ran them on main,
so, I assume they were just broken anyway.

Co-authored-by: czavieh <czavieh@users.noreply.github.com>

* Add -configfile flags to hecklerd and rizzod

This lets us avoid having to copy config files around and will get us
closer to a low-toil integration test.

* Update integration test setup and instructions

This adds more details to the steps needed to even run the tests, plus
it makes them actually work. I decided that rather than having people
copy the sample configs I would make it so that they point to them with
the new -configfile argument, but I may change my mind about it later.
I'd like to get to the point where the integration tests have their own
config for the test app that we use to template the hecklerd and rizzod
configs, rather than making people copy and/or edit them.

* Define integration tests' success criteria somewhat concretely

We also decided to make the commands we ask the user to run a bit
simpler, by offloading the actually running of the daemons to systemd.
Since, for some reason, the foreground processes are not just the
daemons. I could not figure out a way to use build args to make two
images that run one or the other service, so unfortunately, I had to
switch the CMD to an ENTRYPOINT and make the service name an argument to
it. Really not a fan, but, whatever. At least now we are satisfied that
our changes thus far have not broken the app.

Co-authored-by: czavieh <czavieh@users.noreply.github.com>

---------

Co-authored-by: czavieh <czavieh@users.noreply.github.com>