More integration test automation, plus bug fixes (#22)
* Automate the integration tests partway

Setting up the app, the test repo, the conf file, and actually compiling
the binaries are still in their own steps, and we don't have very robust
things going on in the integration tests (or even a way to make it exit
non-0 if they fail) but this still eliminates a fair chunk of tedium.
More improvements to come.

* Fix ssh config Include line

Co-authored-by: czavieh <czavieh@users.noreply.github.com>

* Clean up ~/.ssh/config when running make clean

Since we can't use -o flags for this part, this is the next best thing.

* Move integration tests into a bash script

Also fix known_hosts shenanigans in gen-ssh-keys so that you can run
`make clean` and it will not interfere with your next test run (also,
you don't have to manually approve every ssh connection!)

Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>

* Fix bugs with using non-default branch

The hecklerd process needs to check out the branch you intend to use.
The code path needs to bubble errors up to make that traceable.

Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>
Co-authored-by: czavieh <czavieh@users.noreply.github.com>

* Automatically clean up lingering GH issues before running tests

Requires write perms on repo Administration. We will write up the list
of perms that the test app needs vs the regular heckler app when we are
done working on these changes

Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>
Co-authored-by: czavieh <czavieh@users.noreply.github.com>

* Also check that commit authors are collaborators

We took the time to refactor the previous code for getting the app token
so that it reduces some of the github API header copypasta. Also, that
entire block of code for eliminating lingering issues is its own
function now. We will write any other preflight checks like this.

Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>
Co-authored-by: czavieh <czavieh@users.noreply.github.com>

* Miscellaneous readability, usability, and grammar improvements

Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>
Co-authored-by: czavieh <czavieh@users.noreply.github.com>

* Update integration test documentation again

Fixed formatting and added API permission dependencies

Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>
Co-authored-by: czavieh <czavieh@users.noreply.github.com>

* Add some rudimentary assertions about post-run state

Putting serverspecs aside for now, this lets us make some very basic
assertions that the issues that heckler created are all closed, since
these integration tests run in a way that makes heckler auto-close them
as if they are approved.

Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>
Co-authored-by: czavieh <czavieh@users.noreply.github.com>

* We also need to make sure that we have some Issues in the first place

We were tempted to assert the exact number, but we know that that would
easily break if make-repo was changed, and we'd rather just be more
robust and cross-check the issues and commits someday than bother with
an annoying magic number to have to change if the make-repo script
changes.

Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>

* Fixed a typo

* Add new TODO about testing autoTag behvior

Co-authored-by: czavieh <czavieh@users.noreply.github.com>

---------

Co-authored-by: czavieh <czavieh@users.noreply.github.com>
Co-authored-by: crashdummy101 <129228667+crashdummy101@users.noreply.github.com>