Modernize protoc compilation
Since the old code was using an old, deprecated, non-official library to
generate code with protoc, this took a lot more work than I expected.
The new generated code includes mutexes and other things that shouldn't
be copied by value, and since everything in Golang is copy-by-value,
everything that was passing around a type from the auto-generated code
needed to be refactored. Furthermore, comparison between these types was
no longer possible to do as we had been doing before, since they now
contained complex types that were not compatible with simple `==` or
even the one naughty instance of using `cmp.Equal`. Not even using
`reflect.DeepEqual` would work since some of the nested types were not
comparable and also you would not want to compare them anyway. So I had
to write a `ShallowEqual` that heavily uses the `reflect` library to go
deeper than a true shallow comparison (`==`) but not as deep as
`reflect.DeepEqual`. I'm not 100% convinced I like how it handles
certain cases around polymorphism (as much as you can get that in
Golang) nor how it gets picky about comparing different primitive types
that otherwise do have the same value (easy examples are signed and
unsigned ints with the same actual numerical value, or floats with no
rounding errors that are exactly equal to integers), but it does the job
well enough for our purposes to replace struct comparisons we were doing
before on the simpler struct types.