Do compilation from top-level Makefile instead
paths=source_relative does a lot of work here. It changed the names of
some things internal to the generated code, and required searching for
.proto files, but that seems worth it for the sheer convenience factor.

Also, I updated the README with instructions on how to generate the code
and filled out some stuff near there which seemed to need correcting.