Add more tests for comparing structs
Right now the implementation is such that order matters, not field
names, but we are considering making it so that you can have fields in
different orders as long as they're all the same name. For now, these
tests cover the cases about ordering and naming that are currently true.