Modernize protoc compilation (#23)
* Modernize protoc compilation

Since the old code was using an old, deprecated, non-official library to
generate code with protoc, this took a lot more work than I expected.
The new generated code includes mutexes and other things that shouldn't
be copied by value, and since everything in Golang is copy-by-value,
everything that was passing around a type from the auto-generated code
needed to be refactored. Furthermore, comparison between these types was
no longer possible to do as we had been doing before, since they now
contained complex types that were not compatible with simple `==` or
even the one naughty instance of using `cmp.Equal`. Not even using
`reflect.DeepEqual` would work since some of the nested types were not
comparable and also you would not want to compare them anyway. So I had
to write a `ShallowEqual` that heavily uses the `reflect` library to go
deeper than a true shallow comparison (`==`) but not as deep as
`reflect.DeepEqual`. I'm not 100% convinced I like how it handles
certain cases around polymorphism (as much as you can get that in
Golang) nor how it gets picky about comparing different primitive types
that otherwise do have the same value (easy examples are signed and
unsigned ints with the same actual numerical value, or floats with no
rounding errors that are exactly equal to integers), but it does the job
well enough for our purposes to replace struct comparisons we were doing
before on the simpler struct types.

* Do compilation from top-level Makefile instead

paths=source_relative does a lot of work here. It changed the names of
some things internal to the generated code, and required searching for
.proto files, but that seems worth it for the sheer convenience factor.

Also, I updated the README with instructions on how to generate the code
and filled out some stuff near there which seemed to need correcting.

* Rename "util" to "pbutil"

Short for "protobuff util." Google's style guide discourages generic
names like "util" but says names that are based on them are fine.

* Add more tests for comparing structs

Right now the implementation is such that order matters, not field
names, but we are considering making it so that you can have fields in
different orders as long as they're all the same name. For now, these
tests cover the cases about ordering and naming that are currently true.

* Fix how ShallowEqual works, and document it much better

It's a bit shameful to admit this, but I had absolutely no idea until I
started writing more test specs that Golang doesn't truly have "unnamed"
struct fields. They just get the name of their type if you don't give
them one. Therefore, there's no need to think about cases of two structs
with "unnamed" fields in a different order as being unequal, because if
you gave those fields proper names and made them the same, they'd
basically be the same thing. Anyway, this function now returns true
basically so long as you either have all the same data in the same
order, no matter what you call it, or you have the same data in whatever
order, but with consistent names.

I also decided to write a more proper godoc for this package, even
though we're not generating any yet, simply because it makes sure the
behavior is well-explained.

* Fix typo

* Fix typos and break early in second loop for perf benefits

Also renamed the "ok" variable to something more meaningful