Cache issue numbers to reduce load on Github API
When using github.com, as the repo gets more issues in it, searches
become increasingly expensive API calls. With 100k+ issues, we regularly
get ratelimited when searching for commits. Therefore, this commit adds
simple caching of issue numbers when searching. This means that we only
incur the API cost of searching once. The cache operation fails open, so
if it can't write to disk for whatever reason it won't interrupt the
operation of the loop.

Additionally, we cache the issue number when creating it to further
minimize lookups.